CFG: sort loop targets by variable name
Otherwise, the order depends on whatever Python prefers during a certain
execution (set don't have a specified ordering, tuple({1,2}) can be
(1,2) or (2,1)). This is convenient because the order is re-used as the
order for the arguments of the loop invariants (so we don't have to adapt
the invariant for each run).

Signed-off-by: Mathieu PATUREL <z5312157@unsw.edu.au>