DSA: always insert joiner nodes, even they aren't used later
Because they might be used later in the invariants, which we don't know
about at the DSA stage.

If we want to add this optimization back in, we need to (1) take away the
contexts artifact or (2) sync it up correctly with the graph.

This catches a pretty nasty bug (which I don't think could break
soundness, but it took me a while to debug it):

    int a = 0;
    for (int i = 0; i < 10; i++)
    // invariant: 0 <= i <= 10
    //            i <= 2 ==> a = 0
    {
        if (i == 2)
        {
            a = a + 1;
        }
    }

Apply DSA:

    // inv i a = (0 <= i <= 10) && (i <= 2 ==> a = 0)

    int a1 = 0;
    int i1 = 0;
    // prove $ inv i1 a1
    while (i2 < 10)
    {
        // assume $ inv i2 a2
        if (i2 == 2)
        {
            a3 = a2 + 1;
        }
        i3 = i2 + 1;
        // prove $ inv i3 a4  // !! this is the issue !!
    }

a4, the joiner for a3 a2, is only used in the invariant. So it wasn't
emitted.

This means that we don't know anything about a4, and so I couldn't prove
anything about it (this is why I think it didn't break soundness).

Regardless, this is a scary bug.

Signed-off-by: Mathieu PATUREL <z5312157@unsw.edu.au>