Nip: initialize all #assigned variable to False
Except for function arguments (TODO: and globals)

This technically isn't required, because the SMT solver needs to prove
that every #assigned variable is true on assignment. Without these
initialization, it has to consider both cases where it's true
(redundant) and false. This patch allows the SMT solver to just
consider the cases where it is false.

Signed-off-by: Mathieu PATUREL <z5312157@unsw.edu.au>