Fix weakness in validate_dsa.ensure_assigned_at_most_once
When obtaining a list of assigned variables in a node, we used a helper
function which returned a *set* of assigned variables. When writing the
test originally, this helper function returned a list, and thus we
didn't have this issue. However, in the general case, it makes sense
for the helper function to use a set. This is how the bug crept in.

Note that if this bug had ever occurred, it would've been caught by an
assertion in the DSA transformation, and if a variable had multiple
type, the generated SMT would probably have been invalid. Regardless,
this is bug with soundness breaking potential.

TODO: write report
Signed-off-by: Mathieu PATUREL <z5312157@unsw.edu.au>