Loop invariants for deliverable (verification succeeds)
To verify the handler loop, we need to establish a loop invariant
relating the `Lc_unhandled_notified` and `lc_last_handled_notified` sets of
the platform state to the `lbadge` variable of the handler loop, and the
state of the `have_reply` variable of the handler loop to the constructor of
the `lc_unhandled_reply` variable of the platform state.

Note: there was a slight spec change, `handler_loop_iter_pre` should guarantee
that `lc_last_handled_notified` is empty on loop entry. This condition was
missing from the Dec 5 version of `CoreAbstractSpec`, which stated only that
`lc_unhandled_notified` should be empty on entry.